app/requirements.txt
# Web framework and server
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
httpx>=0.27.0

# Utils
python-dotenv==1.0.1
pydantic==2.11.3
pydantic-settings==2.1.0
PyJWT==2.8.0

# Database
alembic==1.13.1
supabase==2.15.0  # Python client for Supabase

pytest-asyncio
